Core: Refactor C targets for reuse
- Merge crypto implementations into single files
- Merge WireGuard targets
- Externalize logger to partout_log
- Fix tun.h nullability
- Add pp_dns_resolve
- Move pp_file_read to common.c
- Delete pp_log_simple_append
