+++ Sources/PartoutCore_C/common.c
@@ -22,21 +22,85 @@ void pp_clog_v(pp_log_category category,
 #endif
     va_list args;
     va_start(args, fmt);
-    // Add 1 to include the null terminator
-    const size_t msg_len = 1 + vsnprintf(NULL, 0, fmt, args);
+    const int formatted_len = vsnprintf(NULL, 0, fmt, args);
+    va_end(args);
+    if (formatted_len < 0) {
+#if !PARTOUT_WINDOWS
+        errno = saved_errno;
+#endif
+        return;
+    }
+    const size_t msg_len = (size_t)formatted_len + 1;
     char *msg = pp_alloc(msg_len);
+    va_start(args, fmt);
     vsnprintf(msg, msg_len, fmt, args);
     va_end(args);
-    pp_clog(category, level, msg);
+    partout_log(category, level, msg);
     pp_free(msg);
 #if !PARTOUT_WINDOWS
     errno = saved_errno;
 #endif
 }
 
-#if PARTOUT_ANDROID
-#include <android/log.h>
+static FILE *pp_file_open_read(const char *path) {
+#ifdef _WIN32
+    FILE *file = NULL;
+    return (fopen_s(&file, path, "rb") == 0) ? file : NULL;
+#else
+    return fopen(path, "rb");
+#endif
+}
 
+char *pp_file_read(const char *rel_path, const char *parent) {
+    char *abs_path = NULL;
+    FILE *file = NULL;
+    char *buffer = NULL;
+
+    /* Prepend parent if not NULL. */
+    if (parent) {
+        const int path_len = snprintf(NULL, 0, "%s/%s", parent, rel_path);
+        if (path_len < 0) goto failure;
+        abs_path = calloc(1, path_len + 1);
+        if (!abs_path) goto failure;
+        snprintf(abs_path, path_len + 1, "%s/%s", parent, rel_path);
+    } else {
+        abs_path = pp_dup(rel_path);
+        if (!abs_path) goto failure;
+    }
+
+    /* Open file at absolute path. */
+    file = pp_file_open_read(abs_path);
+    if (!file) goto failure;
+    free(abs_path);
+    abs_path = NULL;
+
+    /* Compute file size. */
+    if (fseek(file, 0, SEEK_END) != 0) goto failure;
+    long size = ftell(file);
+    if (size < 0) goto failure;
+    rewind(file);
+
+    /* Allocate buffer (+1 for '\0'). */
+    size_t buffer_size = (size_t)size;
+    if ((long)buffer_size != size) goto failure;
+    if (buffer_size > SIZE_MAX - 1) goto failure;
+    buffer = malloc(buffer_size + 1);
+    if (!buffer) goto failure;
+    size_t read_size = fread(buffer, 1, buffer_size, file);
+    fclose(file);
+    file = NULL;
+
+    if (read_size != buffer_size) goto failure;
+    buffer[buffer_size] = '\0';
+    return buffer;
+failure:
+    if (buffer) free(buffer);
+    if (file) fclose(file);
+    if (abs_path) free(abs_path);
+    return NULL;
+}
+
+#if PARTOUT_ANDROID
 JNIEnv *pp_jni_attach_thread(bool *did_attach) {
     JNIEnv *env;
     jint status = (*jvm)->GetEnv(jvm, (void **)&env, JNI_VERSION_1_6);
@@ -84,42 +148,6 @@ void pp_jni_delete_global_ref(void *ref) {
     PP_JNI_DETACH(env);
 }
 
-void pp_log_simple_append(const char *tag, pp_log_level level, const char *message) {
-    const char *log_tag = tag ? tag : "Partout";
-    int android_level = 0;
-    switch (level) {
-    case PPLogLevelDebug:
-        android_level = ANDROID_LOG_VERBOSE;
-        break;
-    case PPLogLevelInfo:
-        android_level = ANDROID_LOG_DEBUG;
-        break;
-    case PPLogLevelNotice:
-        android_level = ANDROID_LOG_INFO;
-        break;
-    case PPLogLevelError:
-        android_level = ANDROID_LOG_WARN;
-        break;
-    case PPLogLevelFault:
-        android_level = ANDROID_LOG_FATAL;
-        break;
-    }
-    __android_log_print(android_level, log_tag, "%s", message);
-}
-#else
-void pp_log_simple_append(const char *tag, pp_log_level level, const char *message) {
-    FILE *out = NULL;
-    switch (level) {
-    case PPLogLevelError:
-    case PPLogLevelFault:
-        out = stderr;
-        break;
-    default:
-        out = stdout;
-        break;
-    }
-    fprintf(out, "%s[%d]: %s\n", tag ? tag : "Partout", level, message);
-}
 #endif
 
 #if !PARTOUT_WINDOWS
