+++ Sources/PartoutCore_C/include/portable/endian.h
@@ -5,7 +5,7 @@
  */
 
 #pragma once
-#include "portable/conditionals.h"
+#include "conditionals.h"
 
 #include <stdint.h>
