+++ Sources/PartoutCrypto_C/crypto_aead_mbedtls.h
@@ -10,8 +10,8 @@
 #include <stddef.h>
 #include "portable/common.h"
 #include "crypto/crypto_base.h"
-#include "hmac_mbed.h"
-#include "crypto_aead_mbed_api.h"
+#include "hmac_mbedtls.h"
+#include "crypto_aead_mbedtls_api.h"
 
 #define PP_MBED_AEAD_AES_BLOCK_SIZE (size_t)16
 #define PP_MBED_AEAD_GCM_IV_LENGTH (size_t)12
