+++ Sources/PartoutCrypto_C/crypto_mbedtls.h
@@ -10,9 +10,9 @@
 #include <stddef.h>
 #include "portable/common.h"
 #include "crypto/crypto.h"
-#include "crypto_mbed_api.h"
-#include "crypto_aead_mbed.h"
-#include "hmac_mbed.h"
+#include "crypto_mbedtls_api.h"
+#include "crypto_aead_mbedtls.h"
+#include "hmac_mbedtls.h"
 
 #define PP_MBED_AES_BLOCK_SIZE (size_t)16
