+++ Sources/PartoutCrypto_C/crypto_mock.c
