--- Sources/PartoutCrypto_C/openssl/crypto.c
@@ -1,18 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2026 Davide De Rosa
- *
- * SPDX-License-Identifier: GPL-3.0
- */
-
-#include <openssl/rand.h>
-#include "crypto/crypto.h"
-#include "crypto_openssl.h"
-
-bool pp_openssl_crypto_init_seed(const uint8_t *src, const size_t len) {
-    unsigned char x[1];
-    if (RAND_bytes(x, 1) != 1) {
-        return false;
-    }
-    RAND_seed(src, (int)len);
-    return true;
-}
