+++ Sources/PartoutWireGuard/Internal/WireGuardBackend.swift
@@ -2,7 +2,7 @@
 // Copyright © 2018-2023 WireGuard LLC. All Rights Reserved.
 
 internal import _PartoutPortable_C
-internal import _PartoutWireGuardBackend_C
+internal import _PartoutWireGuard_C
 
 /// A enum describing WireGuard log levels defined in `api-apple.go`.
 enum WireGuardLogLevel: Int32 {
