--- Sources/PartoutWireGuardBackend_C/include/module.modulemap
@@ -1,4 +1,0 @@
-module _PartoutWireGuardBackend_C {
-    header "wireguard/backend.h"
-    export *
-}
