+++ Sources/PartoutWireGuard_C/include/wireguard/wireguard.h
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: GPL-3.0
  */
 
+#include "wireguard/backend.h"
 #include "wireguard/key.h"
 #include "wireguard/logging.h"
 #include "wireguard/x25519.h"
